New min-packets

+++ src/NetComm.h
@@ -176,7 +176,7 @@ typedef struct pkt14 {
 
 typedef struct pkt15 {
     uint32_t    data;
-    uint8_t     vel;
+    uint16_t     vel;
 } PKT_MIN_POS_UPDATE;
 #define MIN_10 15
 
@@ -185,7 +185,7 @@ typedef struct pkt16 {
     uint32_t    playersOnFloor;
     uint32_t    xPos[11];
     uint32_t    yPos[11];
-    uint8_t     vel[32];
+    uint16_t     vel[32];
 } PKT_MIN_ALL_POS_UPDATE;
 #define MIN_11 16
